+++ apiphany-core/src/test/java/org/apiphany/client/http/JavaNetHttpExchangeClientTest.java
@@ -727,6 +727,24 @@ class JavaNetHttpExchangeClientTest {
 		}
 
 		@Test
+		void shouldConvertByteArraySupplierToByteArrayBodyPublisher() {
+			ApiClientFluentAdapter request = ApiClientFluentAdapter.of(apiClient)
+					.body(() -> BYTES);
+
+			BodyPublisher bodyPublisher = JavaNetHttpExchangeClient.toBodyPublisher(request);
+
+			assertThat(bodyPublisher.contentLength(), equalTo((long) BYTES.length));
+
+			ByteBufferSubscriber subscriber = new ByteBufferSubscriber();
+			bodyPublisher.subscribe(subscriber);
+			subscriber.awaitCompletion();
+
+			assertTrue(subscriber.isCompleted());
+			assertNull(subscriber.getError());
+			assertThat(BYTES, equalTo(subscriber.getReceivedBytes()));
+		}
+
+		@Test
 		void shouldConvertInputStreamToInputStreamBodyPublisher() {
 			ByteArrayInputStream bis = new ByteArrayInputStream(BYTES);