+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `OneShotInputStreamSupplier` class that allows only a single `get` call, useful for retries.
 - Added `OneShotHttpEntity` for Apache HTTP Client 5 clients that allows the input stream to be consumed only once.
 - Added `ApacheHC5Entities.create(InputStream, ContentType)` similar to Apache `HttpEntities.create` methods but for handling input streams.
+- Added `ApiClientFluentAdapter.body(Supplier)` to allow lazy body building.
 
 ---